Re-render gemspecs to Ruby, rather than copying the original
Gems pulled out of git typically use `git ls-files` to generate a list
of files, but that doesn't work if we strip out the .git
directories. Since we've already loaded the gemspec and already have a
list of files, we can just re-render it back to Ruby and use that.

Fixes #51